pyproject.toml
[tool.poetry]
name = "fate-scheduler"
version = "0.0.1.post1"
description = "The operating system-level command scheduler and manager."
license = "MIT"
repository = "https://github.com/chicago-cdac/fate"
authors = [
  "Jesse London <jesselondon@gmail.com>",
  "Kyle MacMillan <macmillan@uchicago.edu>",
]
packages = [{include = "fate", from = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
#
# FIXME: See dssg/argcmdr#27
#
# argcmdr = "^0.13"
#
argcmdr = {url = "https://github.com/dssg/argcmdr/archive/0.13.0.zip"}
pyyaml = "^6.0"
toml = "^0.10.2"

[tool.poetry.dev-dependencies]

[tool.poetry.scripts]
fate = "fate:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
